add Controller Product and Category
